src/test/java/ca/wendyliu/didemo/controllers/PropertyInjectedControllerTest.java
package ca.wendyliu.didemo.controllers;

import ca.wendyliu.didemo.services.GreetingServiceImpl;
import org.junit.Before;
import org.junit.Test;

import static org.junit.Assert.assertEquals;

/** Test for BAD EXAMPLE (property injection) */
public class PropertyInjectedControllerTest {

    private PropertyInjectedController propertyInjectedController;

    @Before
    public void setUp() throws Exception {
        this.propertyInjectedController = new PropertyInjectedController();
        this.propertyInjectedController.greetingService = new GreetingServiceImpl();
    }

    @Test
    public void testGreeting() throws Exception {
        assertEquals(GreetingServiceImpl.GREETING_MESSAGE, propertyInjectedController.sayHello());
    }
}
